Move parameters into categories
Qpid is probably going into the top-level of the installer with
the removal of puppet-capsule, so we need to ensure the installer
--help isn't growing out of control.  This hides everything but
log-level, the only setting a katello user will really need
to touch.

For non-katello users, this has the benefit of categorizing the
params and making it a bit easier to use.